+++ crates/nostr/src/prelude.rs
@@ -48,6 +48,7 @@ pub use crate::nips::nip17::{self, *};
 pub use crate::nips::nip19::{self, *};
 pub use crate::nips::nip21::{self, *};
 pub use crate::nips::nip22::{self, *};
+#[allow(deprecated)]
 pub use crate::nips::nip26::{self, *};
 pub use crate::nips::nip34::{self, *};
 pub use crate::nips::nip35::{self, *};