+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 
 - nostr: deprecate `nip21::extract_from_text` function (https://github.com/rust-nostr/nostr/pull/923)
 - nostr: deprecate `Tags::from_text` constructor
+- nostr: deprecate NIP-26 (https://github.com/rust-nostr/nostr/pull/928)
 
 ## v0.42.1 - 2025/05/26